api/requirements.txt
arrow==0.13.0
astroid==1.6.0
blinker==1.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.4
gunicorn==19.7.1
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
jsonpickle==0.9.5
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.6
pylint==1.8.1
python-dateutil==2.6.1
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
stripe==1.82.1
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
psycopg2-binary==2.7.4